requirements.txt
newsapi-python==0.2.7
python-dotenv==1.0.1
transformers==4.41.0
torch==2.2.2
pyo==1.0.5
deepfilternet==0.5.6
python-osc==1.8.3
schedule==1.2.2
googletrans==4.0.0-rc1
gdeltdoc
pandas>=2.2.3
requests>=2.32.3
typing_extensions>=4.13.0